Rename and remove unused files
