+++ notifications-kibana-plugin/public/pages/Notifications/component/NotificationsHistogram/HistogramControls.tsx
